+++ model/scratch/tests/test_slr.py
@@ -68,11 +68,10 @@ class TestSLR(AsyncTestCase):
                 res = await mctx.accuracy(sctx)
                 self.assertTrue(0.0 <= res < 1.0)
                 # Test predict
-                async for repo, prediction, confidence in mctx.predict(
-                    sctx.repos()
-                ):
+                async for repo in mctx.predict(sctx.repos()):
                     correct = FEATURE_DATA[int(repo.src_url)][1]
                     # Comparison of correct to prediction to make sure prediction is within a reasonable range
+                    prediction = repo.prediction().value
                     self.assertGreater(prediction, correct - (correct * 0.10))
                     self.assertLess(prediction, correct + (correct * 0.10))
 
@@ -90,9 +89,8 @@ class TestSLR(AsyncTestCase):
     async def test_02_predict(self):
         async with self.sources as sources, self.features as features, self.model as model:
             async with sources() as sctx, model(features) as mctx:
-                async for repo, prediction, confidence in mctx.predict(
-                    sctx.repos()
-                ):
+                async for repo in mctx.predict(sctx.repos()):
                     correct = FEATURE_DATA[int(repo.src_url)][1]
+                    prediction = repo.prediction().value
                     self.assertGreater(prediction, correct - (correct * 0.10))
                     self.assertLess(prediction, correct + (correct * 0.10))